fix(stella-plugin,stella-runtime): reconcile with main, and split the envelope reader out of wire.rs
Two things the rebase onto main surfaced, kept in one commit because both are
consequences of landing this branch on a tree that moved under it rather than
changes of their own.

**`LoopGrant` literals.** `before_turn_stages` (#3543, earlier on this branch)
is a new field, and four hand-built `LoopGrant` values arrived on main while
this branch was open — three test helpers and `host_call.rs`'s own. Each takes
`Vec::new()`, which is the "every stage this program runs" default the field
documents. `permits_hook_is_declared_and_graded_even_on_a_hand_built_grant`
gains the matching assertion for `permits_stage`, since the whole point of that
test is that a hand-built grant below steering leaks nothing, and the stage
filter is a strengthening of the point filter rather than a second door.

**`wire.rs` crossed 1500 lines.** #3518's visitor added roughly ninety lines to
a file that main had grown independently, and a first-time crossing gets no
baseline entry — it gets split. `src/wire/envelope.rs` takes the framing (the
two-key envelope, the field identifier, `FromEnvelope`, `decode_body` and the
visitor); `src/wire.rs` keeps the vocabulary those types arrive in. That is a
real seam rather than a line-count dodge: one file is what a plugin says, the
other is the shape it says it in.

`make guards-fast` passes, file-size included.
