+++ crates/stella-runtime/tests/wrapper_verdict.rs
@@ -474,6 +474,7 @@ fn the_hold_allowance_is_the_ask_clamped_to_the_hosts_ceiling() {
         participation: Participation::Arbiter,
         hooks: vec![stella_plugin::HookEvent::Stop],
         points: vec![stella_plugin::WrapperPoint::AfterTurn],
+        before_turn_stages: Vec::new(),
         calls: Vec::new(),
         max_calls: None,
         max_fanout_width: None,
@@ -553,6 +554,7 @@ fn an_abstention_does_not_buy_another_turn() {
             participation: Participation::Arbiter,
             hooks: vec![stella_plugin::HookEvent::Stop],
             points: vec![stella_plugin::WrapperPoint::AfterTurn],
+            before_turn_stages: Vec::new(),
             calls: Vec::new(),
             max_calls: None,
             max_fanout_width: None,
